laid out number pad screen, but:
cancel button current position is easy to mistap when tapping save button when adding points quickly/blindly which user will likely do as this is the most frequent action in the app and meant to be done quickly. so will move cancel button position and change layout in next commit before developing elsewhere.